PWA-1658 Renamed getPageConfig to fetchPageConfig and added backwards compatibility.
