+++ libraries/common/actions/page/index.js
@@ -1,0 +1,1 @@
+export { default as fetchPageConfig } from './fetchPageConfig';
